# Speed up MasterIndex.Each