What does this PR change? What problem does it solve?
-----------------------------------------------------
`MasterIndex.Each` currently uses channels and along with a goroutine to relay data from `Index.Each` to a unified channel. This PR replaces the channel with a simple callback which drastically reduces the CPU overhead:

```
name                old time/op  new time/op  delta
MasterIndexEach-16   6.68s ±24%   0.96s ± 2%  -85.64%  (p=0.008 n=5+5)
```

Profiling `MasterIndexEach` shows that now most time is spent inside `index.toPackedBlob` which converts the indexEntry to a PackedBlob. It might be possible to make that conversion lazy, although that might just shift some of the costs.

Was the change previously discussed in an issue or on the forum?
----------------------------------------------------------------
The problem showed up in https://github.com/restic/restic/issues/3916 .
<!--
Link issues and relevant forum posts here.

If this PR resolves an issue on GitHub, use "Closes #1234" so that the issue
is closed automatically when this PR is merged.
-->

Checklist
---------

<!--
You do not need to check all the boxes below all at once. Feel free to take
your time and add more commits. If you're done and ready for review, please
check the last box. Enable a checkbox by replacing [ ] with [x].
-->

- [x] I have read the [contribution guidelines](https://github.com/restic/restic/blob/master/CONTRIBUTING.md#providing-patches).
- [x] I have [enabled maintainer edits](https://help.github.com/en/github/collaborating-with-issues-and-pull-requests/allowing-changes-to-a-pull-request-branch-created-from-a-fork).
- [x] I have added tests for all code changes.
- ~~[ ] I have added documentation for relevant changes (in the manual).~~
- [ ] There's a new file in `changelog/unreleased/` that describes the changes for our users (see [template](https://github.com/restic/restic/blob/master/changelog/TEMPLATE)).
- [x] I have run `gofmt` on the code in all commits.
- [x] All commit messages are formatted in the same style as [the other commits in the repo](https://github.com/restic/restic/blob/master/CONTRIBUTING.md#git-commits).
- [x] I'm done! This pull request is ready for review.